+++ app/content/tests/test_tasks.py
@@ -7,6 +7,7 @@ from app.content.factories import EventFactory
 from app.content.tasks.event import (
     run_post_event_actions,
     run_sign_off_deadline_reminder,
+    run_sign_up_start_notifier,
 )
 from app.util.utils import now
 
@@ -159,3 +160,73 @@ def test_post_event_actions_set_runned_post_event_actions_to_true():
     event.refresh_from_db()
 
     assert event.runned_post_event_actions
+
+
+@pytest.mark.django_db
+@patch("app.content.tasks.event.__sign_up_start_notifier")
+def test_sign_up_start_notifier_is_called_when_not_runned_and_time_due(
+    mock_sign_up_start_notifier,
+):
+    """Event sign up start notifier should be runned when not runned and the time is due"""
+    EventFactory()
+    EventFactory(start_registration_at=now())
+
+    run_sign_up_start_notifier()
+
+    assert mock_sign_up_start_notifier.call_count == 2
+
+
+@pytest.mark.django_db
+@patch("app.content.tasks.event.__sign_up_start_notifier")
+def test_sign_up_start_notifier_is_not_called_when_time_not_due(mock_sign_up_start_notifier):
+    """Event sign up start notifier should not be runned when the event ended today or has not ended yet"""
+    EventFactory(start_registration_at=now() + timedelta(minutes=1))
+
+    run_sign_up_start_notifier()
+
+    assert not mock_sign_up_start_notifier.called
+
+
+@pytest.mark.django_db
+@patch("app.content.tasks.event.__sign_up_start_notifier")
+def test_sign_up_start_notifier_is_not_called_when_already_runned(mock_sign_up_start_notifier):
+    """Event sign up start notifier should not be runned when they have already runned"""
+    EventFactory(runned_sign_up_start_notifier=True)
+
+    run_sign_up_start_notifier()
+
+    assert not mock_sign_up_start_notifier.called
+
+
+@pytest.mark.django_db
+@patch("app.content.tasks.event.__sign_up_start_notifier")
+def test_sign_up_start_notifier_is_not_called_when_event_closed(mock_sign_up_start_notifier):
+    """Event sign up start notifier should not be runned when the event is closed"""
+    EventFactory(closed=True)
+
+    run_sign_up_start_notifier()
+
+    assert not mock_sign_up_start_notifier.called
+
+
+@pytest.mark.django_db
+@patch("app.content.tasks.event.__sign_up_start_notifier")
+def test_sign_up_start_notifier_is_not_called_when_not_sign_up(mock_sign_up_start_notifier):
+    """Event sign up start notifier should not be runned when event.sign_up is false"""
+    EventFactory(sign_up=False)
+
+    run_sign_up_start_notifier()
+
+    assert not mock_sign_up_start_notifier.called
+
+
+@pytest.mark.django_db
+def test_sign_up_start_notifier_set_runned_sign_up_start_notifier_to_true():
+    """Event sign up start notifier should set event.runned_sign_up_start_notifier to true"""
+    event = EventFactory()
+
+    run_sign_up_start_notifier()
+
+    event.refresh_from_db()
+
+    assert event.runned_sign_up_start_notifier